patient-service/pyproject.toml
[tool.poetry]
name = "patient-service"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"
packages = [{include = "src/*.py"}]

[tool.poetry.dependencies]
python = "^3.9"
sqlalchemy = "1.4.46"
psycopg2-binary = "^2.9.5"
alembic = "^1.9.4"
black = "^23.1.0"
python-dotenv = "^1.0.0"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
strawberry-graphql = "^0.159.0"

[tool.poetry.scripts]
start = "src.app:start"
generate_migrations = "migrations:generate_migrations"
create_migrations = "migrations:create_migration"
migrate = "migrations:upgrade_migration"
rollback_migration = "migrations:downgrade_migration"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
